requirements.txt
fastapi
sqlalchemy
psycopg2
uvicorn
python-dotenv